iohelper: move some public codes into
