+++ util/compression/compression.go
@@ -5,12 +5,12 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"sync"
 
 	cdcompression "github.com/containerd/containerd/archive/compression"
 	"github.com/containerd/containerd/content"
 	"github.com/containerd/containerd/images"
 	"github.com/containerd/stargz-snapshotter/estargz"
+	"github.com/moby/buildkit/util/iohelper"
 	digest "github.com/opencontainers/go-digest"
 	ocispecs "github.com/opencontainers/image-spec/specs-go/v1"
 	"github.com/pkg/errors"
@@ -253,52 +253,5 @@ func decompress(ctx context.Context, cs content.Store, desc ocispecs.Descriptor)
 			return nil, err
 		}
 	}
-	return &readCloser{r, ra.Close}, nil
-}
-
-type readCloser struct {
-	io.ReadCloser
-	closeFunc func() error
-}
-
-func (rc *readCloser) Close() error {
-	err1 := rc.ReadCloser.Close()
-	err2 := rc.closeFunc()
-	if err1 != nil {
-		return errors.Wrapf(err1, "failed to close: %v", err2)
-	}
-	return err2
-}
-
-type WriteCloser struct {
-	io.WriteCloser
-	CloseFunc func() error
-}
-
-func (wc *WriteCloser) Close() error {
-	err1 := wc.WriteCloser.Close()
-	err2 := wc.CloseFunc()
-	if err1 != nil {
-		return errors.Wrapf(err1, "failed to close: %v", err2)
-	}
-	return err2
-}
-
-type Counter struct {
-	n  int64
-	mu sync.Mutex
-}
-
-func (c *Counter) Write(p []byte) (n int, err error) {
-	c.mu.Lock()
-	c.n += int64(len(p))
-	c.mu.Unlock()
-	return len(p), nil
-}
-
-func (c *Counter) Size() (n int64) {
-	c.mu.Lock()
-	n = c.n
-	c.mu.Unlock()
-	return
+	return &iohelper.ReadCloser{ReadCloser: r, CloseFunc: ra.Close}, nil
 }
